update project setup to use pytests
