+++ setup.cfg
@@ -1,14 +1,5 @@
 [egg_info]
 
-[nosetests]
-with-coverage=1
-cover-package=airspeed
-cover-inclusive=1
-traverse-namespace=1
-detailed-errors=1
-where=tests
-with-doctest=1
-
 [isort]
 profile = black
