+++ setup.py
@@ -28,11 +28,12 @@ setup(
         "cachetools",
     ],
     extras_require={"dev": [
-        "black==22.3.0",
-        "isort==5.12.0",
-        "flake8>=6.0.0",
-        "flake8-black>=0.3.6",
-        "flake8-isort>=6.0.0",
+        "black",
+        "isort",
+        "flake8",
+        "flake8-black",
+        "flake8-isort",
+        "pytest",
     ]},
     test_suite="tests",
     tests_require=[],
